fix bug on query reset=true, add buy city & taxes feature
