+++ src/app/game/components/board/BoardNormal.tsx
@@ -92,8 +92,6 @@ function TileCity({ data }: {data: {[key:string]: string|number}}) {
     // tile data
     type TileCityType = {name: string, price: number, img: string, info: string, square: number}
     const { name, price, img, info, square } = data as TileCityType 
-    // price label
-    const priceText = `after:block after:content-[attr(data-price)]`
     // get room info
     const getGameRoomInfo = gameState.gameRoomInfo.map(v => v.room_id).indexOf(gameState.gameRoomId)
     // set curse
@@ -106,10 +104,53 @@ function TileCity({ data }: {data: {[key:string]: string|number}}) {
                     : `${name};${translateInfo}`
     // highlight
     const isPlayerOnTop = gameState.gamePlayerInfo.map(v => v.pos).indexOf(square)
+    // tile info
+    const tileInfo = name.match(/special/i) ? 'special' : name.match(/cursed/i) ? 'cursed' : 'city'
+    // price label
+    const priceText = `after:block after:content-[attr(data-price)]`
+    // match city with player data
+    let getCityData: any[] = null
+    gameState.gamePlayerInfo.map(player => {
+        const cityList = player.city?.split(';') || []
+        // city name
+        let tempCityName = name.match(/\d/) ? translateUI({lang: miscState.language, text: name as any}) : name
+        // match current city
+        const isCityBought = cityList.map(v => v.split('*')[0]).indexOf(name)
+        if(isCityBought !== -1) {
+            // set city owner
+            tempCityName += `\n${player.display_name}`
+            // check city property (match the latest prop)
+            type CityPropertyType = 'land'|'1house'|'2house'|'2house1hotel'
+            const cityProperty = cityList[isCityBought].match(/2house1hotel$|2house$|1house$|land$/)[0] as CityPropertyType
+            switch(cityProperty) {
+                // [owner, price, property]
+                case 'land': 
+                    getCityData = [tempCityName, player.display_name, price + (price * .10), '1house', '']
+                    return
+                case '1house': 
+                    getCityData = [tempCityName, player.display_name, price + (price * .20), '2house', '🏡']
+                    return
+                case '2house': 
+                    getCityData = [tempCityName, player.display_name, price + (price * .30), '2house1hotel', '🏡🏡']
+                    return
+                case '2house1hotel': 
+                    getCityData = [tempCityName, player.display_name, price + (price * .40), 'realestate', '🏡🏡🏨']
+                    return
+            }
+        }
+        // city not bought
+        else {
+            getCityData = [tempCityName, null, price, 'land', null]
+            return
+        }
+    })
+    const [cityName, cityOwner, cityPrice, cityProperty, cityIcon] = getCityData as [string, string, number, string, string]
+    // city info
+    const cityInfo = cityOwner ? `${name},${cityProperty},${cityPrice},${cityOwner}` : `${name},${cityProperty},${cityPrice}`
 
     return (
         <div className="relative">
-            <div className="absolute z-10" data-player-path={square}>
+            <div className="absolute z-10" data-player-path={square} data-tile-info={tileInfo} data-city-info={cityInfo}>
                 {gameState.gamePlayerInfo.map((player, i) => player.pos == square ? <Characters key={i} playerData={player}/> : null)}
             </div>
             <div data-tooltip={newInfo.replaceAll(';', '\n')} className="relative flex flex-col">
@@ -118,11 +159,9 @@ function TileCity({ data }: {data: {[key:string]: string|number}}) {
                 {/* tile label */}
                 <div className={`${isPlayerOnTop !== -1 ? 'shadow-inner-md shadow-green-400' : ''} 
                 font-mono ml-px w-[7.1vw] h-[6.75vh] bg-darkblue-4/90 text-black text-center`}>
-                    <p className={`leading-3 lg:leading-relaxed text-[2vh] ${priceText}`} data-price={moneyFormat(price)}> 
-                        {name.match(/\d/)
-                            ? translateUI({lang: miscState.language, text: name as any})
-                            : name
-                        } 
+                    <p className={`${cityIcon == '' ? '' : priceText} leading-3 lg:leading-relaxed text-[2vh] whitespace-pre`} 
+                    data-price={moneyFormat(cityPrice)}> 
+                        {cityName} 
                     </p>
                 </div>
             </div>
