+++ src/app/home/HomePage.tsx
@@ -17,8 +17,11 @@ export default function HomePage() {
     useEffect(() => {
         // get url params
         const resetData = location.search.match('reset=true')
-        // reset all data
-        if(resetData) resetAllData(gameState)
+        // reset all data and remove query
+        if(resetData) {
+            history.pushState({}, null, '/')
+            resetAllData(gameState)
+        }
         // check token for auto login
         if(miscState.secret) checkAccessToken(miscState, gameState)
         // navigate to room list
