+++ src/app/room/components/RoomCard.tsx
@@ -1,10 +1,11 @@
 import { FormEvent, useEffect, useState } from "react"
 import { useMisc } from "../../../context/MiscContext"
-import { applyTooltipEvent, fetcher, fetcherOptions, qS, setInputValue, translateUI } from "../../../helper/helper"
-import { ICreateRoom, IGameContext, IShiftRoom, IMiscContext, IResponse } from "../../../helper/types"
+import { applyTooltipEvent, qS, translateUI } from "../../../helper/helper"
+import { ICreateRoom, IGameContext, IMiscContext } from "../../../helper/types"
 import { useGame } from "../../../context/GameContext"
 import Link from "next/link"
 import SelectCharacter from "./SelectCharacter"
+import { deleteRoom, joinRoom, spectateRoom } from "../helper/functions"
 
 export default function RoomCard({ roomData }: {roomData: ICreateRoom['list']}) {
     const miscState = useMisc()
@@ -244,219 +245,5 @@ function manageFormSubmits(ev: FormEvent<HTMLFormElement>, roomId: number, miscS
         case `spectate_button_${roomId}`: spectateRoom(roomId, gameState); break
         // delete room function
         case `delete_button_${roomId}`: deleteRoom(formInputs, roomId, gameState); break
-    }
-}
-
-function spectateRoom(roomId: number, gameState: IGameContext) {
-    // submit buttons
-    const joinButton = qS(`#join_button_${roomId}`) as HTMLButtonElement
-    const spectateButton = qS(`#spectate_button_${roomId}`) as HTMLButtonElement
-    const deleteButton = qS(`#delete_button_${roomId}`) as HTMLButtonElement
-    // disable buttons
-    spectateButton.disabled = true
-    joinButton ? joinButton.disabled = true : null
-    deleteButton ? deleteButton.disabled = true : null
-    // set player as spectator
-    gameState.setSpectator(true)
-    // move to game room
-    const link = qS(`#gotoGame${roomId}`) as HTMLAnchorElement
-    link.click()
-    // enable buttons
-    joinButton.removeAttribute('disabled')
-    spectateButton ? spectateButton.removeAttribute('disabled') : null
-    deleteButton ? deleteButton.removeAttribute('disabled') : null
-}
-
-async function joinRoom(formInputs: HTMLFormControlsCollection, roomId: number, miscState: IMiscContext, gameState: IGameContext) {
-    // submit buttons
-    const joinButton = qS(`#join_button_${roomId}`) as HTMLButtonElement
-    const spectateButton = qS(`#spectate_button_${roomId}`) as HTMLButtonElement
-    const deleteButton = qS(`#delete_button_${roomId}`) as HTMLButtonElement
-    // result message
-    const resultMessage: Element = qS(`#result_room_${roomId}`)
-    // input value container
-    const inputValues: IShiftRoom = {
-        action: 'room join',
-        room_id: roomId.toString(),
-        room_password: null, // from player input
-        confirm_room_password: null, // from input value
-        display_name: gameState.myPlayerInfo.display_name,
-        money_start: null,
-        select_character: null
-    }
-    // get input elements
-    for(let i=0; i<formInputs.length; i++) {
-        const input = formInputs.item(i) as HTMLInputElement
-        if(input.nodeName.match(/INPUT/)) {
-            // filter inputs
-            if(setInputValue('rules', input)) {
-                // get money start
-                const moneyStart = input.value.match(/start: \d+/)[0].split(': ')[1]
-                inputValues.money_start = moneyStart
-            }
-            else if(setInputValue('confirm_room_password', input)) inputValues.confirm_room_password = input.value.trim().toLowerCase()
-            // skip other inputs
-            else if(input.id.match(/room_id|room_name|room_password_\d+|player_count|player_max|creator|select_character/)) continue
-            // error
-            else {
-                if(resultMessage) {
-                    gameState.setRoomError(inputValues.room_id)
-                    setTimeout(() => gameState.setRoomError(null), 2000);
-                    resultMessage.textContent = `${input.id} error`
-                    return
-                }
-                joinButton.classList.add('text-red-600')
-                joinButton.textContent = 'err400'
-                return
-            }
-        }
-    }
-    // set character value
-    const selectCharacter = qS(`#select_character_${roomId}`) as HTMLInputElement
-    inputValues.select_character = selectCharacter.value
-    // matching password
-    const roomPassword = qS(`#room_password_${inputValues.room_id}`) as HTMLInputElement
-    // password doesnt match
-    if(roomPassword.value != '' && roomPassword.value != inputValues.confirm_room_password) {
-        gameState.setRoomError(inputValues.room_id)
-        setTimeout(() => gameState.setRoomError(null), 2000);
-        resultMessage.textContent = translateUI({lang: miscState.language, text: 'wrong password'})
-        return
-    }
-    inputValues.room_password = roomPassword.value == '' ? null : roomPassword.value
-    delete inputValues.confirm_room_password
-    // submit button loading
-    const tempButtonText = joinButton.textContent
-    joinButton.textContent = 'Loading'
-    joinButton.disabled = true
-    // disable other buttons
-    spectateButton ? spectateButton.disabled = true : null
-    deleteButton ? deleteButton.disabled = true : null
-    // fetch
-    const joinRoomFetchOptions = fetcherOptions({method: 'PUT', credentials: true, body: JSON.stringify(inputValues)})
-    const joinRoomResponse: IResponse = await (await fetcher('/room', joinRoomFetchOptions)).json()
-    // response
-    switch(joinRoomResponse.status) {
-        case 200: 
-            // save access token
-            if(joinRoomResponse.data[0].token) {
-                localStorage.setItem('accessToken', joinRoomResponse.data[0].token)
-                delete joinRoomResponse.data[0].token
-            }
-            // set joined room
-            gameState.setMyCurrentGame(roomId)
-            // get room info
-            const findRoomData = gameState.roomList.map(v => v.room_id).indexOf(roomId)
-            const { room_id, room_name, creator, rules } = gameState.roomList[findRoomData]
-            // split rules
-            const splitRules = rules.match(/^board: (normal|delta|2_way);dice: (1|2);start: (50000|75000|100000);lose: (-25000|-50000|-75000);mode: (5_laps|7_laps|survive);curse: (5|10|15)$/)
-            // remove main rules
-            splitRules.splice(0, 1)
-            const [board, dice, money_start, money_lose, mode, curse] = [
-                splitRules[0], // board
-                +splitRules[1], // dice
-                +splitRules[2], // money start
-                +splitRules[3], // money lose
-                splitRules[4], // mode
-                +splitRules[5], // curse
-            ]
-            // set room info & filter to prevent duplicate
-            gameState.setGameRoomInfo(rooms => [...rooms, {
-                room_id, room_name, creator, 
-                board, mode, money_lose, curse, dice
-            }].filter((obj1, i, arr) => 
-                arr.findLastIndex(obj2 => obj2.room_name == obj1.room_name) === i
-            ))
-            // move to game room
-            const link = qS(`#gotoGame${roomId}`) as HTMLAnchorElement
-            link.click()
-            // enable submit buttons
-            joinButton.textContent = tempButtonText
-            joinButton.removeAttribute('disabled')
-            spectateButton ? spectateButton.removeAttribute('disabled') : null
-            deleteButton ? deleteButton.removeAttribute('disabled') : null
-            return
-        default: 
-            // error message
-            gameState.setRoomError(inputValues.room_id)
-            setTimeout(() => gameState.setRoomError(null), 3000);
-            const translateError = translateUI({lang: miscState.language, text: joinRoomResponse.message as any})
-            resultMessage.textContent = `${joinRoomResponse.status}: ${translateError || joinRoomResponse.message}`
-            // enable submit buttons
-            joinButton.textContent = tempButtonText
-            joinButton.removeAttribute('disabled')
-            spectateButton ? spectateButton.removeAttribute('disabled') : null
-            deleteButton ? deleteButton.removeAttribute('disabled') : null
-            // set join button type back to BUTTON
-            joinButton.type = 'button'
-            return
-    }
-}
-
-async function deleteRoom(formInputs: HTMLFormControlsCollection, roomId: number, gameState: IGameContext) {
-    // submit buttons
-    const joinButton = qS(`#join_button_${roomId}`) as HTMLButtonElement
-    const spectateButton = qS(`#spectate_button_${roomId}`) as HTMLButtonElement
-    const deleteButton = qS(`#delete_button_${roomId}`) as HTMLButtonElement
-    // input value container
-    const inputValues = {
-        action: 'room hard delete',
-        display_name: gameState.myPlayerInfo.display_name,
-        creator: null,
-        room_id: null,
-        room_name: null
-    }
-    // get input elements
-    for(let i=0; i<formInputs.length; i++) {
-        const input = formInputs.item(i) as HTMLInputElement
-        if(input.nodeName.match(/INPUT/)) {
-            // filter inputs
-            if(setInputValue('room_id', input)) inputValues.room_id = input.value.trim().toLowerCase()
-            else if(setInputValue('room_name', input)) inputValues.room_name = input.value.trim().toLowerCase()
-            else if(setInputValue('creator', input)) inputValues.creator = input.value.trim().toLowerCase()
-            // skip other inputs
-            else if(input.id.match(/room_password|player_count|player_max|rules|select_character/)) continue
-            // error
-            else {
-                deleteButton.classList.add('text-red-600')
-                deleteButton.textContent = 'err400'
-                return
-            }
-        }
-    }
-    // confirm delete room dialog
-    if(!confirm(`Are you sure wanna delete "${inputValues.room_name}"?`)) return
-    // submit button loading
-    const tempButtonText = deleteButton.textContent
-    deleteButton.textContent = 'Loading'
-    deleteButton.disabled = true
-    // disable other buttons
-    spectateButton ? spectateButton.disabled = true : null
-    joinButton ? joinButton.disabled = true : null
-    // fetch
-    const deleteRoomFetchOptions = fetcherOptions({method: 'DELETE', credentials: true, body: JSON.stringify(inputValues)})
-    const deleteRoomResponse: IResponse = await (await fetcher('/room', deleteRoomFetchOptions)).json()
-    // response
-    switch(deleteRoomResponse.status) {
-        case 200: 
-            // save access token
-            if(deleteRoomResponse.data[0].token) {
-                localStorage.setItem('accessToken', deleteRoomResponse.data[0].token)
-                delete deleteRoomResponse.data[0].token
-            }
-            // enable submit buttons
-            deleteButton.textContent = tempButtonText
-            deleteButton.removeAttribute('disabled')
-            spectateButton ? spectateButton.removeAttribute('disabled') : null
-            joinButton ? joinButton.removeAttribute('disabled') : null
-            return
-        default: 
-            deleteButton.classList.add('text-red-600')
-            deleteButton.textContent = `err${deleteRoomResponse.status}`
-            // enable submit buttons
-            deleteButton.removeAttribute('disabled')
-            spectateButton ? spectateButton.removeAttribute('disabled') : null
-            joinButton ? joinButton.removeAttribute('disabled') : null
-            return
     }
 }
