+++ tsconfig.json
@@ -20,7 +20,8 @@
       {
         "name": "next"
       }
-    ]
+    ],
+    "downlevelIteration": true
   },
   "include": [
     "next-env.d.ts",
